run-jsc-stress-tests should be less verbose https://bugs.webkit.org/show_bug.cgi?id=122397
Reviewed by Sam Weinig.

It now just prints something like:

 965/1831 ....

Where the first number is the number of tests that completed, the second number is the
number of tests total, and the dots are the number of tests running right now - which is
useful towards the end when you're waiting for the last slow tests to finish.

It will also print out the output from failing tests and it will print a line for skipped
tests (to annoy you into getting them into a shape where they can be unskipped). For
example, for LayoutTests, it will print just the diff, if the test fails. It's cool.

* Scripts/run-jsc-stress-tests:



Identifier: 140460@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@156983 268f45cc-cd09-0410-ab3c-d52691b4dbfc