-webkit-cross-fade paints SVGs at full opacity during cross-fade https://bugs.webkit.org/show_bug.cgi?id=122441 <rdar://problem/13973162>
Reviewed by Simon Fraser.

Test: css3/images/cross-fade-svg-with-opacity.html

* platform/graphics/CrossfadeGeneratedImage.cpp:
(WebCore::drawCrossfadeSubimage):
(WebCore::CrossfadeGeneratedImage::drawCrossfade):
Factor the image painting out into a function.
Conditionally use a transparency layer around painting if the image
we're drawing is an SVG (SVGImage is the only Image subclass that will
not respect the context's opacity).

Add a test that ensures that -cross-fade with SVG images draws correctly.

* css3/images/cross-fade-svg-with-opacity-expected.html: Added.
* css3/images/cross-fade-svg-with-opacity.html: Added.


Identifier: 140516@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157045 268f45cc-cd09-0410-ab3c-d52691b4dbfc