Include misspelling dot gap width when centering misspelling dots https://bugs.webkit.org/show_bug.cgi?id=122365
Patch by Myles C. Maxfield <mmaxfield@apple.com> on 2013-10-23
Reviewed by Simon Fraser.

Source/WebCore:

When calculating where to place the misspelling dots, we find the
maximum number of full dots that can fit under the misspelled word,
and then center a run of that many dots. However, when we're
centering the run, we are only considering the size of the extra
fractional dot that we cut off. However, the dot image has a "gap"
of transparent pixels (which visually provide tracking for the dots)
which visually appears to be empty space. We should take this gap
space into consideration when centering the run of dots. We also
should make sure that our dots start on integral pixel boundaries
because otherwise we can't set the phase of the dot run properly.

Test: editing/spelling/centering-misspelling-dots.html

* platform/graphics/mac/GraphicsContextMac.mm:
(WebCore::GraphicsContext::drawLineForDocumentMarker):

LayoutTests:

Created a pixel new test where taking this gap width into account when
performing the centering calculation actually changes the location of
where the dots end up. In addition, created a pixel test that misspells
two words next to each other to make sure that the misspelling dots don't
appear under the space between the words.

* editing/spelling/centering-misspelling-dots-expected.txt: Added.
* editing/spelling/centering-misspelling-dots.html: Added.
* platform/mac/editing/spelling/centering-misspelling-dots-expected.png: Added.
* editing/spelling/misspelling-dots-dont-extend-beyond-words-expected.txt: Added.
* editing/spelling/misspelling-dots-dont-extend-beyond-words.html: Added.
* platform/mac/editing/spelling/misspelling-dots-dont-extend-beyond-words-expected.png: Added.

Identifier: 141343@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157900 268f45cc-cd09-0410-ab3c-d52691b4dbfc